sw: FP64 reduce weight and weight gradient precision to float

+++ sw/applications/include/network.h
@@ -65,7 +65,7 @@ typedef struct network_fp64_t_ {
     uint32_t OUT_CH;
 
     double *b;
-    double *W;
+    float *W;
     double *b_grad;
     double *W_grad;
